Resolve guest-bios review findings
Split oversized orchestration and guest-bios test modules so the file-level
pylint suppressions are no longer needed. Keep the public generation module as
the re-export surface while moving routing and orchestration internals into
focused private modules.

Tighten the verified guest-bios review issues: use enum route keys, narrow the
subprocess lint suppression, and count guest-bios TEI blocks through parsed XML.
Update the guest-bios ADR and docs references to use a unique ADR identifier
and add screen-reader context for XML examples.
