+++ docs/execplans/2-3-3-generate-guest-bios-from-reference-document-bindings.md
@@ -26,9 +26,6 @@ service can use Vidai Mock as the inference provider, retrieve guest profiles
 through existing binding resolution, and produce deterministic TEI body
 enrichment.
 
-This is a pre-implementation plan. Do not implement the feature until this
-ExecPlan has been explicitly approved.
-
 ## Constraints
 
 - Preserve hexagonal architecture boundaries from the
@@ -156,7 +153,7 @@ ExecPlan has been explicitly approved.
   prompt builder, and TEI enrichment helper. Focused test command
   `uv run pytest tests/test_guest_bios.py -q` now passes with 5 tests.
 - [x] 2026-05-14: Added
-  `docs/adr/adr-009-guest-bios-tei-representation.md` to record the
+  `docs/adr/adr-010-guest-bios-tei-representation.md` to record the
   `<div type="guest-bios"><list><item corresp="...">` representation.
 - [x] 2026-05-14: Ran `coderabbit review --agent` for milestone 1. It raised
   a missing expected-revision parser test and an ADR footnote wrapping concern;
@@ -299,7 +296,7 @@ parser/emitter dependency.
 Read `docs/adr/adr-004-show-notes-tei-representation.md`,
 `docs/episodic-podcast-generation-system-design.md`, and the relevant
 `tei_rapporteur` body payload expectations. Add a new ADR at
-`docs/adr/adr-009-guest-bios-tei-representation.md`.
+`docs/adr/adr-010-guest-bios-tei-representation.md`.
 
 Write fail-first unit tests in a new `tests/test_guest_bios.py` covering the
 target body shape:
